requirements.txt
langchain==1.2.0
python-dotenv==1.2.1
streamlit==1.37.1
unstructured==0.18.21
tiktoken==0.12.0
faiss-cpu==1.13.1
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
OpenAI==2.13.0
langchain_openai==1.1.5
langchain_classic==1.0.0
langchain_text_splitters==1.1.0
langchain_community==0.4.1